requirements.txt
python-dotenv>=1.0.1
openai>=1.43.0
langgraph==1.0.0
langchain-core>=0.2.29
requests>=2.32.3
pyodbc>=5.1.0
pydantic>=2.9.0
sqlalchemy>=2.0.20